created req's.txt and csv & pkl output files; fixed kwargs bug
